requirements.txt
GitPython==3.1.40
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
groq==0.4.1
requests==2.31.0
